Removed theme context
