+++ src/App.js
@@ -1,7 +1,6 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import './style.css';
 import TodoApp from './components/TodoApp';
-import { ThemeContext } from './context/ThemeContext';
 
 const version = Number.parseFloat(localStorage.getItem('version') || 0);
 const newVersion = 1.2;
@@ -12,23 +11,7 @@ if (version < newVersion) {
 }
 
 localStorage.setItem('version', JSON.stringify(newVersion));
-const savedTheme = localStorage.getItem('theme');
 
 export default function App() {
-	const [theme, setTheme] = useState(savedTheme || 'dark');
-
-	useEffect(() => {
-		document.documentElement.setAttribute('data-theme', theme);
-		localStorage.setItem('theme', theme);
-	}, [theme]);
-
-	const toggleTheme = () => {
-		setTheme(prev => (prev === 'dark' ? 'light' : 'dark'));
-	};
-
-	return (
-		<ThemeContext.Provider value={theme}>
-			<TodoApp toggleTheme={toggleTheme} />
-		</ThemeContext.Provider>
-	);
+	return <TodoApp />;
 }
