+++ src/components/AddTodo.js
@@ -1,16 +1,25 @@
-import React, { useContext, useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import InputBox from './InputBox';
 import './css/addtodo.css';
-import clsx from 'clsx';
 import { FaMoon as DarkThemeIcon, FaSun as LightThemeIcon } from 'react-icons/fa';
-import { ThemeContext } from '../context/ThemeContext';
 
-const AddTodo = ({ addTask, onEditCancel, editTask, saveEditTask, changeTheme }) => {
+const savedTheme = localStorage.getItem('theme');
+
+const AddTodo = ({ addTask, onEditCancel, editTask, saveEditTask }) => {
 	const [title, setTitle] = useState(editTask ? editTask.title : '');
 	const [content, setContent] = useState(editTask ? editTask.content : '');
 	const [errors, setErrors] = useState({ title: false, content: false });
 
-	const theme = useContext(ThemeContext);
+	const [theme, setTheme] = useState(savedTheme || 'dark');
+
+	useEffect(() => {
+		document.documentElement.setAttribute('data-theme', theme);
+		localStorage.setItem('theme', theme);
+	}, [theme]);
+
+	const changeTheme = () => {
+		setTheme(prev => (prev === 'dark' ? 'light' : 'dark'));
+	};
 
 	const onChangeTitle = e => {
 		setTitle(e.target.value);
@@ -43,34 +52,29 @@ const AddTodo = ({ addTask, onEditCancel, editTask, saveEditTask, changeTheme })
 		}
 		setErrors({ title: title.length === 0, content: content.length === 0 });
 	};
-	const taskTracker = clsx('task-tracker', theme);
-	const btn = clsx('btn', theme);
-	const themeBtn = clsx('theme-button', theme);
-	const edit = clsx('btn edit margin', theme);
-	const remove = clsx('btn cancel', theme);
 	return (
 		<div className="add-todo-container" onKeyPress={ev => ev.key === 'Enter' && addTodo()}>
 			<div className="task-tracker-container">
-				<h2 className={taskTracker}>Task Tracker</h2>
+				<h2 className="task-tracker">Task Tracker</h2>
 				<div className="btns-container-add">
 					{!editTask && (
-						<button className={themeBtn} onClick={changeTheme} aria-label="Switch theme">
+						<button className="theme-button" onClick={changeTheme} aria-label="Switch theme">
 							{theme == 'dark' ? <LightThemeIcon /> : <DarkThemeIcon />}
 						</button>
 					)}
 
 					{!editTask && (
-						<button className={btn} onClick={addTodo}>
+						<button className="btn" onClick={addTodo}>
 							Add
 						</button>
 					)}
 				</div>
 				{editTask && (
 					<div>
-						<button className={edit} onClick={() => addEditedTodo(title, content)}>
+						<button className="btn edit margin" onClick={() => addEditedTodo(title, content)}>
 							Edit
 						</button>
-						<button className={remove} onClick={onCancel}>
+						<button className="btn cancel" onClick={onCancel}>
 							Cancel
 						</button>
 					</div>
