+++ src/components/Button.js
@@ -1,13 +1,13 @@
-import React from "react";
-import clsx from "clsx";
+import React from 'react';
+import clsx from 'clsx';
 
 const Button = ({ children, outline, className, ...rest }) => {
-  const css = clsx("button", className, outline && "outline");
-  return (
-    <button {...rest} className={css}>
-      {children}
-    </button>
-  );
+	const css = clsx('button', className, outline && 'outline');
+	return (
+		<button {...rest} className={css}>
+			{children}
+		</button>
+	);
 };
 
 export default Button;
