+++ src/components/InputBox.js
@@ -1,19 +1,15 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import './css/inputbox.css';
 import clsx from 'clsx';
-import { ThemeContext } from '../context/ThemeContext';
 
 const InputBox = ({ label, value, onChange, hasError }) => {
-	const theme = useContext(ThemeContext);
 	const css = clsx('title-input input', hasError && 'required');
-	const textBox = clsx('text-box', theme);
-	const labelBox = clsx('label-box', theme);
 	return (
 		<div className={css}>
-			<label htmlFor={label} className={labelBox}>
+			<label htmlFor={label} className="label-box">
 				{label}
 			</label>
-			<input id={label} type="text" className={textBox} value={value} onChange={onChange} />
+			<input id={label} type="text" className="text-box" value={value} onChange={onChange} />
 		</div>
 	);
 };
