+++ src/components/Task.js
@@ -1,23 +1,19 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import Button from './Button';
 import clsx from 'clsx';
 import './css/task.css';
-import { ThemeContext } from '../context/ThemeContext';
 
 const Task = ({ task, onRemove, showCompleted, onEdit }) => {
-	const theme = useContext(ThemeContext);
-	const completed = task.completed ? 'completed' : '';
-	const title = clsx('task-title', theme);
-	const edit = clsx('edit', theme);
+	const completed = clsx('task-container', task.completed && 'completed');
 	return (
 		<div className={`task-container ${completed}`}>
 			<div className="task-content" onClick={() => showCompleted(task.id)}>
-				<h3 className={`${title} ${completed}`}>{task.title}</h3>
-				<p className={title}>{task.content}</p>
+				<h3 className={`task-title ${completed}`}>{task.title}</h3>
+				<p className="task-title">{task.content}</p>
 			</div>
 			<div>
 				<Button
-					className={edit}
+					className="edit"
 					onClick={() => {
 						onEdit(task);
 					}}
