+++ src/components/css/addtodo.css
@@ -1,6 +1,7 @@
 .add-todo-container {
 	margin-bottom: 60px;
 }
+
 .todo-create-container {
 	margin: 10px 0;
 }
@@ -46,9 +47,11 @@
 .edit {
 	border: 1px solid #3ec300;
 }
+
 .edit:hover {
 	background: #3ec300;
 }
+
 .margin {
 	margin-left: 5px;
 }
@@ -57,16 +60,19 @@
 	border-radius: 4px;
 	padding: 3px;
 }
+
 .theme-button {
 	background-color: transparent;
 	outline: none;
 	border: none;
 	font-size: 20px;
-	color: black;
-}
-.theme-button.dark {
 	color: yellow;
 }
+
+html[data-theme='light'] .theme-button {
+	color: black;
+}
+
 .btns-container-add {
 	display: flex;
 	align-items: center;
