+++ src/components/css/task.css
@@ -2,19 +2,23 @@
 	border-bottom: 1px solid var(--inputHover);
 	margin-bottom: 20px;
 }
+
 .task-container {
 	display: flex;
 	justify-content: space-between;
 	align-items: center;
 	user-select: none;
 }
+
 .task-title {
 	color: var(--textColor);
 }
+
 .task-content {
 	overflow: hidden;
 	width: 960px;
 }
+
 .task-title.completed {
 	color: green;
 }
@@ -61,6 +65,7 @@
 		width: 100%;
 	}
 }
+
 @media only screen and (max-width: 1200px) {
 	.task-content {
 		max-width: 80%;
