+++ src/components/css/todoapp.css
@@ -7,6 +7,7 @@
 	padding: 28px;
 	border-radius: 5px;
 }
+
 @media only screen and (max-width: 600px) {
 	.container {
 		border: none;
