--- src/context/ThemeContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const ThemeContext = createContext('dark');
-
-export { ThemeContext };
